+++ components/bottombar/impl/src/main/java/com/flipperdevices/bottombar/impl/composable/bottombar/ComposeBottomBar.kt
@@ -30,6 +30,7 @@ import androidx.compose.ui.unit.dp
 import com.flipperdevices.bottombar.impl.model.FlipperBottomTab
 import com.flipperdevices.bottombar.impl.viewmodel.BottomBarViewModel
 import com.flipperdevices.connection.api.ConnectionApi
+import com.flipperdevices.core.ui.ktx.SetUpNavigationBarColor
 import com.flipperdevices.core.ui.ktx.tab.tabIndicatorOffset
 import com.flipperdevices.core.ui.theme.LocalPallet
 import tangle.viewmodel.compose.tangleViewModel
@@ -109,6 +110,9 @@ private fun ComposeBottomBarInternal(
             }
         }
     }
+    SetUpNavigationBarColor(LocalPallet.current.bottomBarBackground)
+
+    connectionApi.CheckAndShowUnsupportedDialog()
 }
 
 @Composable